[dtos] move dtos to each domain category
