+++ src/main/java/com/example/threehealthymeals/api/MenuApi.java
@@ -3,8 +3,8 @@ package com.example.threehealthymeals.api;
 import com.example.threehealthymeals.domain.restaurant.Menu;
 import com.example.threehealthymeals.domain.restaurant.MenuRepository;
 import com.example.threehealthymeals.service.MenuService;
-import com.example.threehealthymeals.web.dto.MenuCreateRequest;
-import com.example.threehealthymeals.web.dto.MenuResponse;
+import com.example.threehealthymeals.web.dto.menu.MenuCreateRequest;
+import com.example.threehealthymeals.web.dto.menu.MenuResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
