+++ src/main/java/com/example/threehealthymeals/service/MenuService.java
@@ -5,8 +5,8 @@ import com.example.threehealthymeals.domain.food.FoodRepository;
 import com.example.threehealthymeals.domain.restaurant.Menu;
 import com.example.threehealthymeals.domain.restaurant.MenuRepository;
 import com.example.threehealthymeals.domain.restaurant.RestaurantRepository;
-import com.example.threehealthymeals.web.dto.MenuCreateRequest;
-import com.example.threehealthymeals.web.dto.MenuResponse;
+import com.example.threehealthymeals.web.dto.menu.MenuCreateRequest;
+import com.example.threehealthymeals.web.dto.menu.MenuResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
