+++ src/main/java/com/example/threehealthymeals/web/dto/menu/MenuResponse.java
@@ -1,4 +1,4 @@
-package com.example.threehealthymeals.web.dto;
+package com.example.threehealthymeals.web.dto.menu;
 
 import com.example.threehealthymeals.domain.restaurant.Menu;
 import lombok.Getter;
