+++ src/main/java/com/example/threehealthymeals/web/dto/restaurant/RestaurantListRequest.java
@@ -1,4 +1,4 @@
-package com.example.threehealthymeals.web.dto;
+package com.example.threehealthymeals.web.dto.restaurant;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
